refactor: mpp payment viewer usecase
