+++ wallet-sdk-ui/src/iosMain/kotlin/com/michaeltchuang/walletsdk/ui/liquidStream/IosViewerPaymentOrchestrator.kt
@@ -341,7 +341,7 @@ class IosViewerPaymentOrchestrator(
 
     // ── Signing helpers ───────────────────────────────────────────────────────
 
-    private suspend fun signClaimMessage(
+    suspend fun signClaimMessage(
         claimMessage: ByteArray,
         signer: MppWalletSigner,
         address: String,
